improved code based on feedback
