+++ Samples/InpaintHTTP/Program.cs
@@ -4,6 +4,8 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Threading;
+using Nancy;
+using Nancy.Hosting.Self;
 
 namespace InpaintHTTP
 {
@@ -12,8 +14,12 @@ namespace InpaintHTTP
         static void Main(string[] args)
         {
             Console.WriteLine("Creating new server..");
-            WebHost webHost = new WebHost();
-            Thread.Sleep(-1);
+            NancyHost nancyHost = new NancyHost(new Uri("http://localhost:8069"));
+            nancyHost.Start();
+            Console.WriteLine("Initialized and started NancyHost");
+
+            //inf sleep to avoid main thread from finishing/terminating
+            Thread.Sleep(-1); 
         }
     }
 }
