+++ Samples/InpaintHTTP/TestWebsite/index.html
@@ -1,18 +1,24 @@
-<center>
-	<h1>Local Inpainting Example</h1>
-	<form enctype="multipart/form-data" action="http://localhost:8069/api/inpaint" method="POST">
-		<input type="hidden" name="MAX_FILE_SIZE" value="100000" />
-			Image: <input name="Image" type="file" /><br />
-			Mask: <input name="Mask" type="file" /><br />
-		<input type="submit" value="Upload File" />
-	</form>
+<!DOCTYPE HTML>
+<html>
+<head>
+	<title>Inpainting API test</title>
+</head>
+<body>
 	<center>
-	<h1>Pulic Inpainting Example</h1>
-	<form enctype="multipart/form-data" action="https://inpaint.ferib.dev/api/inpaint" method="POST">
-		<input type="hidden" name="MAX_FILE_SIZE" value="100000" />
-			Image: <input name="Image" type="file" /><br />
-			Mask: <input name="Mask" type="file" /><br />
-		<input type="submit" value="Upload File" />
-	</form>
-</center>
-</center>
+		<h1>Local Inpainting Example</h1>
+		<form enctype="multipart/form-data" action="http://localhost:8069/api/inpaint" method="POST">
+			<input type="hidden" name="MAX_FILE_SIZE" value="100000" />
+				Image: <input name="Image" type="file" /><br />
+				Mask: <input name="Mask" type="file" /><br />
+			<input type="submit" value="Upload File" />
+		</form>
+		
+		<h1>Pulic Inpainting Example</h1>
+		<form enctype="multipart/form-data" action="https://inpaint.ferib.dev/api/inpaint" method="POST">
+			<input type="hidden" name="MAX_FILE_SIZE" value="100000" />
+				Image: <input name="Image" type="file" /><br />
+				Mask: <input name="Mask" type="file" /><br />
+			<input type="submit" value="Upload File" />
+		</form>
+	</center>
+</body>
